add some validation pipeline changes, fix some bugs with auth and profile service

+++ Services/ProfileService/ProfileService.Application/Models/UserProfileInformation.cs
@@ -3,5 +3,5 @@ namespace ProfileService.Application.Models;
 public class UserProfileInformation
 {
     public string NickName { get; set; }
-    public ulong Rating { get; set; }
+    public int Rating { get; set; }
 }
