+++ Services/ProfileService/ProfileService.Application/Queries/GetUserProfileQuery/ProfileResponse.cs
@@ -5,6 +5,6 @@ public class ProfileResponse
     public string NickName { get; init; } = string.Empty;
     public string Email { get; init; } = string.Empty;
     public int Age { get; init; }
-    public ulong Rating { get; init; }
+    public int Rating { get; init; }
     public List<string> FriendIds { get; init; } = new();
 }
