+++ Services/ProfileService/ProfileService.Migrations/Migrations/0001_CreateUserProfile.cs
@@ -14,7 +14,7 @@ public class CreateUserProfile : Migration
                 nick_name TEXT NOT NULL,
                 email TEXT NOT NULL,
                 age int,
-                rating bigint,
+                rating int,
                 friends text,
                 CONSTRAINT EMAIL_UNIQUE_CHECK UNIQUE (email),
                 CONSTRAINT RATING_POSITIVE CHECK (rating >= 0));
